Separated GameServer and CardServer and added game state in GameServer
